requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
yt-dlp==2024.1.21
moviepy==1.0.3
SpeechRecognition==3.10.0
pydub==0.25.1
google-generative-ai==0.4.1